requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
altair==5.0.1
    # via streamlit
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
bcrypt==4.0.1
    # via streamlit-authenticator
blinker==1.6.2
    # via streamlit
cachetools==5.3.1
    # via streamlit
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via streamlit
decorator==5.1.1
    # via validators
diskcache==5.6.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
extra-streamlit-components==0.1.56
    # via
    #   -r requirements.in
    #   streamlit-authenticator
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
humanize==4.6.0
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.6.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
logzero==1.7.0
    # via -r requirements.in
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
newsapi-python==0.2.7
    # via -r requirements.in
numpy==1.24.3
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==0.27.7
    # via -r requirements.in
openpyxl==3.1.2
    # via -r requirements.in
packaging==23.1
    # via streamlit
pandas==2.0.2
    # via
    #   altair
    #   streamlit
pillow==9.5.0
    # via streamlit
protobuf==4.23.2
    # via streamlit
pyarrow==12.0.0
    # via streamlit
pydantic==1.10.9
    # via -r requirements.in
pydeck==0.8.1b0
    # via streamlit
pygments==2.15.1
    # via rich
pyjwt==2.7.0
    # via streamlit-authenticator
pympler==1.0.1
    # via streamlit
pypdfium2==4.15.0
    # via -r requirements.in
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   -r requirements.in
    #   streamlit-authenticator
requests==2.31.0
    # via
    #   newsapi-python
    #   openai
    #   streamlit
rich==13.4.1
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
streamlit==1.23.1
    # via
    #   -r requirements.in
    #   extra-streamlit-components
    #   streamlit-authenticator
streamlit-authenticator==0.2.2
    # via -r requirements.in
tenacity==8.2.2
    # via streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.2
    # via streamlit
tqdm==4.65.0
    # via openai
typing-extensions==4.6.2
    # via
    #   altair
    #   pydantic
    #   streamlit
tzdata==2023.3
    # via
    #   -r requirements.in
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3
    # via streamlit
urllib3==2.0.2
    # via requests
validators==0.20.0
    # via streamlit
watchdog==3.0.0
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-metadata
